+++ config/test.exs
@@ -1,17 +1,22 @@
 use Mix.Config
 
+#config :pigeon, http2_client: Pigeon.Http2.Client.Chatterbox
+
 config :pigeon, :test,
-  valid_gcm_reg_id: System.get_env("VALID_GCM_REG_ID"),
+  fcm_key: System.get_env("GCM_KEY"),
+  valid_fcm_reg_id: System.get_env("VALID_GCM_REG_ID"),
   valid_apns_token: System.get_env("VALID_APNS_TOKEN"),
   apns_cert: "cert.pem",
   apns_key: "key_unencrypted.pem",
   apns_topic: System.get_env("APNS_TOPIC")
 
-config :pigeon, :gcm,
-  key: System.get_env("GCM_KEY")
+config :pigeon, :fcm,
+  fcm_default: %{
+    key: System.get_env("GCM_KEY")
+  }
 
 config :pigeon, :apns,
-  default: %{
+  apns_default: %{
     cert: "cert.pem",
     key: "key_unencrypted.pem",
     mode: :dev