+++ lib/pigeon/apns/config.ex
@@ -1,9 +1,9 @@
 defmodule Pigeon.APNS.Config do
   @moduledoc """
-    Validates configuration settings that initialize APNSWorkers.
+  Validates configuration settings that initialize APNS.Workers.
   """
 
-  def default_name, do: :default
+  def default_name, do: :apns_default
 
   def config(name) do
     config = Application.get_env(:pigeon, :apns)[name]